doals cleanup
