+++ src/sentry/static/sentry/app/components/panels/panelTable.tsx
@@ -73,8 +73,12 @@ const PanelTable = ({
   emptyMessage = t('There are no items to display'),
   loader,
 }: Props) => {
+  const isChildrenValid =
+    React.Children.toArray(children).filter(child => child).length > 0 &&
+    typeof children !== 'object';
+
   const shouldShowLoading = isLoading === true;
-  const shouldShowEmptyMessage = !shouldShowLoading && isEmpty;
+  const shouldShowEmptyMessage = !shouldShowLoading && (isEmpty || !isChildrenValid);
   const shouldShowContent = !shouldShowLoading && !shouldShowEmptyMessage;
 
   return (
