+++ src/sentry/static/sentry/app/views/teamsDashboard/teamDetails/header/modalEditAvatar.tsx
@@ -18,6 +18,7 @@ type Props = {
 type State = {
   isSaving: boolean;
 };
+
 class ModalEditAvatar extends React.Component<Props, State> {
   state: State = {
     isSaving: false,
